On branch full_deployment
 Changes to be committed:
	new file:   getBranches.js
	modified:   index.js
	modified:   package.json
	modified:   updateStageVariables.js

Added getBranches module to use the currentBranch name as the alias for lambda, and to log the name
of the current branch/alias and the rest of the branches/aliases for this lambda. Made version
argument a boolean, since no input is required and added the option using -v instead of -version.