On branch master
 Changes to be committed:
	modified:   index.js
	modified:   package.json
	new file:   test/verifyCorrectAlias.js
	new file:   verifyCorrectAlias.js

Added verifyCorrectAlias with its tests to verify that the lambdaAlias in package.json is the
same as the branch name, as a precaution measure for linking the branch/lambda alias with the
wrong API GW stage. In case of simple upload, only a warning appears. In case of version
upload, an error is thrown. Added lambdaAlias attribute in package.json for testing.